requirements-arm.txt
# HowTimeFly 核心依赖

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0

# 图片处理
pillow>=10.0.0
piexif>=1.1.3

# 视频处理 (可选，需要系统安装 ffmpeg)
ffmpeg-python>=0.2.0

# 配置文件
pyyaml>=6.0

# 视频缩略图 (可选)
opencv-python>=4.8.0
